add actual js script into html
The JS script has to be inside the HTML because the application has to
be self sustaining and won't be served by a webserver.
The index.html will be copied to the machine or run from the local
network.
In this constellation, loading "external" JS via script tag with src
attribute will not work.
The index.js will be kept, because developing and testing locally is
easier with a dedicated JS file.